+++ Dominio/src/main/java/com/ciberdix/th/repositories/VTercerosFamiliaresRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
  * Created by felip on 2/03/2017.
  */
 public interface VTercerosFamiliaresRepository extends CrudRepository<VTercerosFamiliaresEntity, Integer> {
-    List<VTercerosFamiliaresEntity> findByIdTercero(int idTercero);
+    List<VTercerosFamiliaresEntity> findByIdTercero(Long idTercero);
 }
